Refactor the Objective-C integration
