+++ src/win32.mak
@@ -149,7 +149,7 @@ FRONT_SRCS=access.d aggregate.d aliasthis.d apply.d argtypes.d arrayop.d	\
 	dtemplate.d dversion.d entity.d errors.d escape.d			\
 	expression.d func.d globals.d hdrgen.d id.d identifier.d imphint.d	\
 	impcnvtab.d init.d inline.d intrange.d json.d lexer.d lib.d link.d	\
-	mars.d mtype.d nogc.d nspace.d objc_stubs.d opover.d optimize.d parse.d	\
+	mars.d mtype.d nogc.d nspace.d opover.d optimize.d parse.d	\
 	sapply.d sideeffect.d statement.d staticassert.d target.d tokens.d	\
 	traits.d utf.d visitor.d libomf.d scanomf.d typinf.d \
 	libmscoff.d scanmscoff.d
@@ -161,7 +161,7 @@ DMD_SRCS=$(FRONT_SRCS) $(GLUE_SRCS)
 # Glue layer
 GLUEOBJ=glue.obj msc.obj s2ir.obj e2ir.obj tocsym.obj \
 	toobj.obj tocvdebug.obj toir.obj \
-	iasm.obj objc_glue_stubs.obj
+	iasm.obj objc_glue.obj
 
 # D back end
 BACKOBJ= go.obj gdag.obj gother.obj gflow.obj gloop.obj var.obj el.obj \
@@ -186,16 +186,16 @@ SRCS = aggregate.h aliasthis.h arraytypes.h	\
 	attrib.h complex_t.h cond.h ctfe.h ctfe.h declaration.h dsymbol.h	\
 	enum.h errors.h expression.h globals.h hdrgen.h identifier.h idgen.d	\
 	import.h init.h intrange.h json.h lexer.h lib.h macro.h	\
-	mars.h module.h mtype.h nspace.h objc.h parse.h                         \
+	mars.h module.h mtype.h nspace.h parse.h                         \
 	scope.h statement.h staticassert.h target.h template.h tokens.h	\
-	version.h visitor.h objc.d $(DMD_SRCS)
+	version.h visitor.h $(DMD_SRCS)
 
 # Glue layer
 GLUESRC= glue.c msc.c s2ir.c e2ir.c tocsym.c \
 	toobj.c tocvdebug.c toir.h toir.c \
 	irstate.h iasm.c \
 	toelfdebug.d libelf.d scanelf.d libmach.d scanmach.d \
-	tk.c eh.c objc_glue_stubs.c objc_glue.c \
+	tk.c eh.c objc_glue.c \
 	$(GLUE_SRCS)
 
 # D back end
@@ -520,6 +520,9 @@ newman.obj : $(CH) $C\newman.c
 nteh.obj : $C\rtlsym.h $C\nteh.c
 	$(CC) -c $(MFLAGS) $C\nteh
 
+objc_glue.obj : $C\dt.h objc_glue.c
+	$(CC) -c -I$(ROOT) $(MFLAGS) objc_glue
+
 os.obj : $C\os.c
 	$(CC) -c $(MFLAGS) $C\os
 
@@ -595,5 +598,3 @@ ldfpu.obj : vcbuild\ldfpu.asm
 ############################## Generated Rules ###############################
 
 # These rules were generated by makedep, but are not currently maintained
-
-objc_glue_stubs.obj : objc.h objc_glue_stubs.c
